Sketch/prototype of runtime contract types.
Not really a finished change, but something to start the conversation
with; this adds just enough to Serializable and Template instances to
be able to walk their types and auto-generate basic interfaces in
typescript. Without a change like this all that information is still
present, but only available at compile-time for typechecking; the very
limited reflection available in typescript appears not to be adequate
for generating generic functions on Template or Choices.