Move type information out of Serializable and into a "module.meta" entry
This breaks the dependency on a modified @daml/types library, introduces
a dependency on a new @daml/types-meta library, and better separates
concerns. Either this (with the addition of a templateType: Meta field
on templates) or a similar daml-to-json program would be adequate to
provide navigator-like functionality in a @daml/react app, which has
been proven out in an in-progress app.